Move some code to new StudentRegSanityController
There were some functions in the lottery assignment controller that didn't
have much to do with the lottery but were stil useful.  I moved them into
a new controller that can be used similarly to the lottery (instantiate
with a program and any options, then call sanitize()).  I also made the
minor tweaks so the code would run.