Tweak lottery assignment stats test
We were getting incorrect results for the student screwedness value in
some cases.  This was because some students were marked as Interested in
some of the same classes marked as Priority/1.  While we discourage this,
it is possible that some real students will encounter this situation so I
think the test should accommodate it.