Add initial data fixture with US Zip codes
Chapters have been having a problem where they can't use the
location-based user search until someone manually loads Zip codes into
their database.  I'm not sure this is the best solution since it is a big
text file and takes some time (15--30 sec?) to load at migration time, but
I would like to put it in and see if others think it is a good idea.