Add a /learn/[program]/[instance]/catalog_pdf view
This is essentially a redirect to the /manage/ coursecatalog view, except
with the same caching stuff that the normal catalog has.

It depends on the proxy cache; if there are reliability issues with this
we could just save the result in a static file and return the file if it
is not stale.