Allow classrooms to be imported with complete availability
The classroom import form used to match the timeslots from the previous
program with the timeslots of the new program, and make the new classrooms
available only at those timeslots.  Now you can still do that, but there
is a checkbox on the classroom import form to simply create classrooms for
all of the new program's timeslots.