Completion certs: Allow >1 page of descriptions.
Old method: Set zero margins in order to fit letterhead on the page.
Then enclose class descriptions in a minibox to give them margins.

Side effect: If class descriptions go over a page, LaTeX fails to
break the minibox across multiple pages, cutting off all but the first
four or so classes.

New method: Use a different hack to fit the letterhead on the page,
and set sane margins. (The 0.2-inch difference is what our 0.5cm top
margin used to be.)

[Other notes]
* The template program/modules/programprintables/refund_receipt.tex
  may be affected, but it seems not adversely; and we haven't used it
  since MIT was using OMARS for payments.
* I don't understand why \noindent wasn't necessary before. I hope it's
  not relevant.