Remove anchors from AnnouncementLink and Entry
These models were used in some templates to control when prominent
announcements appeared.  We had anchors so that announcements could be
posted to individual programs and classes, but this feature was never
used (to my knowledge).  So, I removed the anchor and made the migration
remove any announcements that weren't "public" (i.e. anchored to Q/Web).
A couple of views and templates have been modified to exclude other uses
of announcements.