Make ForeignKeys to the DataTree allow null values
As part of the transition away from the DataTree, we are going to be
saving model instances without DataTree anchors.  In order for this to
work, all of our anchor fields must allow NULL values.

This raises an interesting question, which is whether these migrations
should be reversible.  In order for them to be reversible we'd have to
write some fairly interesting data migrations.  So far we have made them
irreversible.