Change the way Media objects are associated with programs/classes
Per-class and per-program document uploads were associated with the
class/program through matching DataTree anchors.  This has to go, so I
added a generic field on Media that is restricted to point to only
ClassSubject or Program.  The class documents features of teacher
registration and the AdminMaterials module have been updated to reflect
the change.