Add Printer and PrintRequest models
We used to use UserBits as a means of generating and consuming requests
for printing student schedules on-site.  The Permission and Record models
that now replace UserBit aren't a good fit for this.  This change creates
two simple models:
- Printer, for listing the printers that are available on site
  Note that if none are created, there is effectively one 'global'
  printer.
- PrintRequest, for denoting when a student's schedule should be printed

The onsite (program module) views have been updated to use the new models.