Change signature of our render_to_response()
This is a change with widespread effects.

We had overridden Django's render_to_response() function with our own in
order to sneak more information into the context.  I think this is still
merited, but there was a variable argument feature that really bothered
me.  You could provide just two arguments (template and context) or four
(template, request, program/anchor, and context).  In the latter mode, the
'program/anchor" could consist of a program, a Datatree anchor, or a tuple
of a program and "teach/learn" section name.

Besides referencing the Datatree, this was quite confusing due to the
inconsistent usage; most of our views were just making up something useless
to pass in for the 3rd argument, often incurring an extra query.  It's
going to break a few things but I think it is worth it to remove this
"feature" given the opportunity.

The render_to_response() function now accepts three fixed arguments:
template, request, and context.  There are still some other keyword
arguments.