Add deadline exception for role-specific permissions
UserBits could be assigned to individual users or all users.  So, a
deadline like 'V/Deadline/Registration/Teacher' would technically apply to
everyone including the students.

Now, we have a permission system that allows a deadline to be opened for
particular user types (by setting Permission.role).  We aren't quite ready
to use this yet, because it can result in some odd error messages (like
"teacher registration is not open" instead of "you are not a teacher").

So, for now, ignore deadline check failures when the only reason for
failure is the wrong user role.