Use 'datetime.datetime.now()' in Users migrations
It looks like South is set up to assume that the datetime module has been imported.
In a couple of migrations we were doing a "from datetime import datetime" which
conflicts with this.  The easiest solution is to harmonize with what South is
expecting.

Thanks to Eric for identifying the problem and providing the solution in 5e1fd83.