Fix Program.teachers()['taught_before'] with new schema
Now we can't represent whether a teacher has taught before using a
Q object.  Maybe this feature should be removed... but for now it is
at least partly useful (it selects teachers of previous programs,
and you have to chain it with a query to select teachers of the
current program).