Change the default behavior of Program.getTimeSlots()
Now that we have teacher training and interviews in the mix, there are a
lot of time slot types we usually don't want.  In fact, usually we only
want the class time blocks.

This commit changes Program.getTimeSlots() so that the default behavior is
to filter only class time blocks, but you can provide a list of time slot
types to include or exclude.  The associated Program.getTimeSlotList() and
its invocations are similarly updated.