Add tests for accounting related things in studentreg
Given the new accounting system, it makes sense to increase our test
coverage from zero.

Two new tests within StudentRegTest, one for financial aid (varying the
parameters of financial aid requests and grants) and one for student extra
costs (varying the selected extra items).  The basic point is to check
that we are charging students the right amount of money.