Replace errant colons in QSD names during migrations
This is a minor hack to compensate for some messed up QSD data that MIT
has (and maybe some others).  If a QSD was named something like
"teach:pagename" and anchored under Q/Programs, it would get correctly
renamed to "teach/pagename".  However, if it had been anchored somewhere
else (like Q/Web), that would have been nonsensical under the old scheme;
in our migration, the URL would not have been updated.

This change avoids the problem by substituting a slash into the URL when
it detects a colon right after the "teach/learn[/...]" part of the name.