Show login page when accessing a view using @admin_required
This may be controversial since we generally don't want to give outsiders
any hints, but I'm worried that admins find it confusing when they click a
link and get "permission denied" with no explanation.  This change updates
the @admin_required decorator to behave similar to the program modules
@needs_student, @needs_admin etc. so that a non-logged in user will see a
login page.

If the wrong type of user is logged in, they'll just get PermissionDenied.