Make render_qsd tags expect literal URLs to be quoted
There is a family of tags {% inline[_program]_qsd_block %} and
{% render_inline[_program]_qsd %} that use InlineQSDNode.  This commit
enforces usage consistency of those tags by requiring that the 'url'
argument is either a quoted string, or the name of a template variable.
This avoids the confusing case of an unquoted string that fails to resolve
as a template variable and is then interpreted as a literal.

See discussion of 8d417fd.