Add dynamic open_class_category support to lottery frontend
We now allow an open_class_category Tag to specify the ID of the
ClassCategories that should be used to represent "open" or "walk-in" type
classes for each program.  The lottery now displays the name of that
category and checks all classes against it, instead of using the hardcoded
name "Walk-in Seminars" and assuming that all classes with category symbol
W are in that category.