Re-add variables.less
At one time we were generating a LESS file containing all of the
configuration needed for Bootstrap to work, thus removing the need
for the variables.less file included in Bootstrap.  Now, we are
allowing themes to define a set of variables that they want to
customize.  This set of variables is disjoint from the Bootstrap
variables.  Thus we need to include the Bootstrap variables.