Fix error in ESPUser.__init__()
I saw this on one of my VM dev servers with Ubuntu 13.04.  It said
that the value assigned to a method __name__ attribute had to be
a string, and it was possible for a Unicode object to be passed in.
I'm not sure why we didn't see this before, but the change should
be pretty inconsequential.