Limits accounting to registered students
Optimizes the accounting modules by limiting searches to students who
are registered for the program, rather than all students who have a
Transfer for that program.

In theory this should not change anything, but there exists some bug
that is creating Transfer objects for all users. This bug was crippling
the accounting module before this commit.