Avoid duplicating executed transfers
The accounting controller has a function add_required_transfers() which
essentially initializes a student's financials for a program, by creating
transfers from "accounts receivable" to the program for any required line
item types.  The transfers are deleted and re-created to ensure that they
have up-to-date line items, e.g. if the costs of the program are changed.

If this function is called for a student whose transfers have
been executed, either by submitting a payment or via an administrative
action, then more transfers would be created.  This change should avoid
the problem by not creating any new transfers for line item types that
have already-executed transfers associated with them.