Remove unnecessary datetime.now() statements
StudentRegistration now uses ExpirableModel.valid_objects() and
.is_valid_qobject(), which call datetime.now() internally. This cleans
up the datetime.now() calls that are no longer being used.

We need to keep the datetime.now() in ClassSection.clearStudents(),
and pass 'now' to valid_objects(), because 'now' is used later in the
same function and it's important that both values of now need to be
the same.