Remove selenium tests from .travis.yml
It may be possible to get Selenium tests working in the future, but it's a nontrivial task as they are not directly supported by Travis.  I would like our Travis builds to complete successfully once we've got the kinks out of the main test suite.  It may be a while before the Selenium tests are working.