Customize the CSRF failure page.
Uses the CSRF_FAILURE_VIEW setting to specify a custom view for
rendering the CSRF failure page.

The old failure page was ugly and confusing because it didn't use the
same theme of the website. The custom view uses a template that inherits
from main.html, so theming is maintained even when a CSRF error occurs.

The new template also includes important debugging information that a
user can copy-and-paste into an e-mail to us.