Adds code to fill in lower priorities.
If students haven't filled in their lower choices, we want to make some
guesses so they don't get screwed.  This makes the lottery work more
reasonably when we're using both several priority levels and interested,
and we don't expect all students to fill out all of both.  Details about
the exact conditions are in a comment.