Replaces cls.teachers with cls.teachers.all
There might be more bugs like this, but I can't do this just with grep,
because in some cases cls is a ClassSubject and cls.teachers needs to be
replaced by cls.teachers.all, and in other cases cls is a dictionary and
cls.teachers is an iterable of the teachers for that class. We should
either test all of the other printables that uses cls.teachers, or
manually look at the code for all of those view functions to determine
which other templates need fixing.