Tweak top tab link size/spacing
This was to ensure that the top tabs appeared to align with the body
area.  The underlying problem is that different browsers and platforms
use different fonts, and the different fonts have different baseline
spacing which affects how much the actual link (with its text) is
shifted downwards to align the baseline of the text with the bottom
of the div - we compensate for this, but the amount of shift needed
is platform dependent.

If anyone has more CSS know-how I would appreciate a more
robust solution.