Fix inline QSD editing for blocks using default content
When a QSD does not exist, a new one is created so that we have an
ID to pass to the JSON views if it gets edited.  However, our code
was only saving the new QSD if the user argument was provided,
which is not always true.  By un-indenting "new_qsd.save()", we
ensure that the new QSD object has an ID.