Add django-filebrowser to dependencies
This is a 3rd party extension to the Django admin interface which
allows the uploading of files to an arbitrary range of directories
on the server (e.g. /media/*).  I propose that we use it to allow
admins to modify their images and styles without the need for file
sharing.