Modifications to the mailer cron script.
The script will no longer halt on exceptions. Instead, it will send the exceptions to the sender address and the configured bounce address. It will attempt to retry sending to recipients who it failed to deliver to the first time (this only happens the next time the cron script is executed).