Adds more tests for tags.
In particular, adds tests for tag defaulting and caching behavior, and
in particular for getProgramTag and getBooleanTag.

Note that the caching test is set up to ignore the bug of Issue #866;
when/if that issue is closed, the marked lines may be removed.  This
isn't an ideal solution, but I think it's better than committing a test
that will be failing for the near-term future (and thus is useless),
since I don't want to assume that a deep argcache bug will be fixed
soon.