Changes getProgramTag's defaulting to improve performance.
Once the underlying issues with argcache were fixed, getProgramTag's
defaulting behavior actually made more sense as something closer to the
way it was, since getTag returns None if and only if the tag was not
found.  Passing None as a default rather than passing our default down
therefore has the same behavior, but improves caching performance.