Split up resource module views
This change creates several new functions in the ResourceModule to
handle the functionality specific to certain models (e.g. timeslots,
classrooms, resource types).  This simplifies the main resources()
view function and makes it easier to test the model-specific functions.
This was recommended during the review of pull request #718.

Also creates a ResourceController, which we can use to make the code
cleaner following the upcoming resources schema upgrades.