Allow optional variables to be set in theme editor
In the fruitsalad theme, if N is the number of side tabs, there are really
3N colors to deal with because of the accents (the top tabs, and the left
edge of the side tabs).  The theme accepts N colors as advanced options in
the theme editor, and would calculate the "accent" colors by just
darkening the original by 20% or 30%.

In some cases this isn't enough (for example, if you want to have a black
side tab with a red accent).  To facilitate this, I made the extra 2N
colors "optional" variables.  Instead of displaying all of them on the
theme editor, there is a dropdown box that allows you to override
individual variables.