Used hashed URL as identifier in client side QSD editing
We used to identify QSDs by ID, and now after schema simplification,
we mainly identify them by URL.  (This is helpful in our inline QSD
block rendering, where we don't want to create a QSD object unless the
default text is edited by an admin... so we don't have an ID to work
with.)  The problem is you can't name an HTML DOM element after a URL.

To fix this, pass around URLs instead of IDs in our Javascript, and
compute an 8-character hash of the URL to use as a "slug" for DOM
IDs.

This is another method of addressing the underlying problem that
I (incorrectly) hacked around in ca061ad.