Calculate QSD edit ID only in Python
Instead of having functions in both Javascript and Python to hash the
QSD url, we can just pass it into Javascript in the template.  This
requires passing both the url and the edit_id around on the JS side,
but it lets us remove some unnecessary code.

Thanks to Anthony for pointing this out.