Removes user from InlineQSDNode
The InlineQSDNode class and the inline_qsd_block tag used to take a
template.Variable parameter called user. On the rendered page, the
inline QSD would be editable if and only if the resolved user had the
required permissions.

A while ago, we stopped doing this at the template level, and started
doing this at the CSS/JS level. But we didn't remove the user parameter
from InlineQSDNode / inline_qsd_block, even though it mostly doesn't do
anything anymore.

This commit removes the user parameter, and updates templates to remove
the remaining user arguments.

At the same time, this commit also simplifies some templates (by
replacing trivial uses of inline_qsd_block with the render_inline_qsd
tag), fixes a typo, makes some minor improvements to
InlineQSDNode.render(), and removes the unused QSDStringNode class.

NOTE: This commit changes the behavior of new inline QSDs. Previously, a
QSD object was created and saved on the first load of a page, using the
resolution of the user parameter as the author. Now, since InlineQSDNode
no longer has a user parameter, this will not occur, and the QSD object
will be created on first edit. This may or may not have caching
implications. If it does, the old functionality can easily be restored,
by getting request.user from the context parameter to
InlineQSDNode.render(), and using that user as the author.