Fixes ajax request bug with debug_toolbar
The Django debug toolbar should not be included in all requests, only
page loads. The best way I could think of to accomplish this was by
having SHOW_TOOLBAR_CALLBACK return False if request.is_ajax() is True.

This may have been causing multiple bugs. The bug I noticed was with
inline QSD editing. In esp/public/media/scripts/qsd.js, in
qsd_inline_update(), the function call JSON.parse(data) was failing
because data.content included, in addition to the QSD content, the
content of the Django debug toolbar. So whenever inline QSD was edited,
the new content would be saved correctly, but it would not render the
updated content because qsd_inline_update() was failing.