Changes theme name comment
Changes the theme name comment to be a Django template comment at the
bottom of the main.html template.

The theme name cannot be HTML-commented at the top of the main.html
template, because the XML/HTML declarations must be at the top of the
rendered HTML file. It cannot be HTML-commented at the bottom, because
this is outside all template blocks. And it cannot be Django
template-commented at the top, because {% extends elements/html %} must
be the first template tag.