Moves qsd media includes to elements/html
Now the QSD scripts and stylesheets will be loaded once on all pages.

CSS stylesheets should not be linked outside of the HTML head, and it is
unnecessary to include Javascript files more than once (which could
happen on pages with multiple inline QSDs). Furthermore, it is
negligibly more expensive to load the QSD media files on the few pages
that do not use QSD.