Fixes bug in profile form
The profile form as an optional gender field, which is Tag-disabled by
default. However, the code was using the dictionary bracket accessor to
get the gender data, which fails when the key is not present (which is
the default). Fixes this by changing the access to use get() with a
default of None.