Adds esp.wsgi to repository
Previously, esp.wsgi was always constructed by the setup script. But all
of them should, in theory, be the same, except for the $BASEDIR and the
USE_PROFILER variable. But the base directory can be determined by
python at runtime, and the USE_PROFILER variable can be made into a
Django setting. This commit makes both of these changes, and removes the
creation of esp.wsgi from the setup scripts.