Adds class flags interface to editclass
Only adds it to the manage version, although updates the checking for
that to always include /manage///editclass/#### rather than requiring a
GET or POST as the dashboard buttons produce.  (This is basically for
convenience for those of us who type the URL manually.)  Also some minor
fixes to other parts of flags.