Adds LESS stylesheet to be included by all themes
Overrides Bootstrap styles, or otherwises sets global (to all themes)
styles.

In this particular case, overrides a Bootstrap style that was causing
radio and checkbox labels to appear a line below the input buttons,
rather than inline. This was effecting the rendering of the commpanel
and a few other pages.