Fix SCRMI admin panel slowness
The SCRMI detail view in the admin panel was very slow to load. This
was because it was rendering a dropdown containing all the program
modules in the database, which issues a separate query for each
program module's __str__ method.

Fix this by making SCRMI.module non-editable, so that the dropdown
isn't rendered, since we aren't going to be editing that from the
admin panel anyway.