Default theme configs in get_template_settings
Uses default theme config values in get_template_settings(). Also adds
a default value for fruitsalad's front_page_style.

Previously, the template settings were only those specified in the
theme_template_control Tag. Now, they also include the defaults of any
configs that may not be saved in that Tag.

When a new config is added to the repository, it will not be saved in
the Tag until the next time the theme is recompiled. Until that time,
the template variable will resolve to ''. For some settings, this can
be bad. For example, if front_page_style == '', then index.html will try
including a template whose name is ''. Since this template does not
exist, the include will fail, and the front page will be completely
blank until the next recompile. This commit avoids this problem by
providing for a safe default.