Remove esp.lib.markdownaddons and the associated QSD file-upload.
The QSD page editor used to have a feature where, if you're editing a
page that has a link starting with qsdmedia/ that doesn't resolve to
anything, it allows you to upload a file to that location.

This feature was broken in 17db658464fbdc because the file upload
infrastructure changed and so the files now go to a different
location from where the links point.

Removing it because it is ill-documented, counterintuitive, and
unnecessary bloat. A better way of uploading files from the QSD editor
may be implemented in the future.