Fix bug in displaying comm panel user filters
The "grade", "school" and "grad year" options had disappeared from
the user list selector.  The problem was a hard coded check for
user type which compared against "students" and "teachers", whereas
the user types present on the form are actually "Student" and
"Teacher".  With this change, a somewhat more tolerant comparison
is used to decide which filter options to show.