Store program modules in cache as ProgramModuleObj only
With Django 1.6, we can no longer hack the __class__ attribute
of a model before pickling it.  In order to allow caching of
objects that include program modules, the ProgramModuleObj
methods that return modules have been given a 'convert'
argument that determines whether the ProgramModuleObj instances
are converted to the desired subclass (e.g. AvailabilityModule)
before returning.  The conversion is handled in the new
ProgramModuleObj methods toBaseClass() and toHandlerClass().

This could potentially be fixed by converting program modules
to be proxy models rather than abstract, as Anthony mentioned
in the discussion of #618.