Fix some of the errors in Ajax scheduling backend tests
The tests were written when the change log was stored in memory.
Since then, the system has changed so that the change log is stored
in the database (AJAXChangeLog model) and is now accessed by index
rather than by time (see 04f6854).  This commit updates the tests to
match the new change log scheme.