Preserve Travis memcached install
Looking at recent Travis logs, during the installation of our apt
packages we were upgrading the built-in install of memcached, and
the upgraded version was not started properly (which prevented the
test suite from running).  This change filters memcached out of the
packages to be installed by apt-get, so that we don't touch the
existing install on the Travis VM.

Also removes postgresql from "services:" block, since according to
Travis documentation, postgresql is started at boot.