Add migration to remove leading spaces from QSDs
Before 6f7d69e11136f0b769e47a9c468477d2c58c4cf1, using the inline QSD
editor would add extraneous leading spaces to the content. This is a
bug that has since been fixed, but the leading spaces are still
there. With the new version of Markdown, these leading spaces are
misinterpreted as denoting code blocks, which is undesirable.

Add a migration that, assuming that any leading space is
unintentional, removes it.