Update ThemeController for backup/restore of individual files
This includes functions to check for modifications to the files
(images and scripts) that would normally be copied in by the theme.
The idea is that you ask the user which of the files they would like
to overwrite and which they would like to keep.  For the files they
want to keep, copy them to temporary locations before loading the
theme, then copy them back from the temporary locations.